requirements.txt
addict==2.4.0
fiftyone==0.24.1
gradio==4.36.1
huggingface_hub==0.19.4
ipdb==0.13.13
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.74
Pillow==10.3.0
pycocotools==2.0.7
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
setuptools==59.6.0
setuptools==58.0.4
supervision==0.21.0
termcolor==2.4.0
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.40.1
typer==0.12.3
yapf==0.40.2
